Removed unnecessary PixelRowDelegateProcessor type
